feat(rds): add database + username input vars
This adds some new optional input variables:

 - `database`: the database name
 - `username`: the username used for auth

Both of these values are set in RDS using `name`, which is a nice default but not especially flexible.
It puts a lot of stuttering into the DSN and can be a little verbose when your top-level identifier is namespaced already.

Using a name like `lambda-warehouses` leads to an output `addr` that looks like:

```
postgres://lambda-warehouses:<PASS>@<HOST>/lambda-warehouses
```

(not that `<HOST>` will also contain `lambda-warehouses`, making it appear a total of 3 times.

This PR allows something a bit more clean like:

```
postgres://segment:<PASS>@<HOST>/public
```

Admittedly, this is mostly aesthetic, but I don't think it's unreasonable to want to set this configuration yourself.
(and perhaps there are non-aesthetic use-cases I'm just not thinking of)