feat(vpc): Optionally use NAT instances instead of NAT gateways
IF NAT instances are used, optional inputs are whether to use Elastic
IPs, an SSH key name, and the EC2 instance type. The latest Amazon VPC
NAT AMI is used. The use of NAT instances can be less expensive than NAT
gateways for development VPCs. NAT instances can be stopped when not in
use.